requirements.txt
# Core ML/DL libraries
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
catboost>=1.0.0
shap>=0.41.0

# Data processing and visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# API and web requests
requests>=2.26.0

# Configuration and serialization
pyyaml>=6.0
joblib>=1.1.0

# Logging and utilities
python-json-logger>=2.0.0
tqdm>=4.62.0

# Development tools
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910 